Removendo os arquivos err.h e sys.h
