+++ include/assert/gl_ok.h
@@ -1,0 +1,16 @@
+#ifndef S3D_ASSERT_GL_OK_H 
+#define S3D_ASSERT_GL_OK_H 
+
+#include<GL/GLee.h>
+
+#include "error/abort_because.h"
+
+#define assert_gl_ok() ({ \
+    int err = glGetError(); \
+    \
+    if(err) { \
+        abort_because("OpenGL error: %d.", err); \
+    } \
+})
+
+#endif
