+++ include/file/read_file.h
@@ -1,0 +1,6 @@
+#ifndef S3D_FILE_READ_FILE_H
+#define S3D_FILE_READ_FILE_H
+
+char *s3d_read_file(const char *path, int text, int *size);
+
+#endif
