+++ include/time/fsleep.h
@@ -1,0 +1,6 @@
+#ifndef S3D_TIME_FSLEEP_H
+#define S3D_TIME_FSLEEP_H
+
+void s3d_fsleep(float seconds);
+
+#endif
